# swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Txing AI API文档",
        "title": "Txing AI API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/websites": {
            "post": {
                "description": "创建新的网站",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站管理"
                ],
                "summary": "创建网站",
                "parameters": [
                    {
                        "description": "网站信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateWebsiteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.WebsiteVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/websites/favicon": {
            "post": {
                "description": "自动获取网站的favicon图标",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站管理"
                ],
                "summary": "获取网站图标",
                "parameters": [
                    {
                        "description": "网站地址",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetFaviconReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.GetFaviconVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/websites/list": {
            "get": {
                "description": "获取网站列表，支持分页和搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站管理"
                ],
                "summary": "获取网站列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式(asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "网站名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/websites/{id}": {
            "get": {
                "description": "根据ID获取网站详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站管理"
                ],
                "summary": "获取网站详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "网站ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.WebsiteVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新网站信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站管理"
                ],
                "summary": "更新网站",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "网站ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "网站信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateWebsiteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.WebsiteVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除网站",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站管理"
                ],
                "summary": "删除网站",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "网站ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/api/websites/list": {
            "get": {
                "description": "用户端获取启用状态的网站列表，支持搜索和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "网站"
                ],
                "summary": "用户端获取网站列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "网站名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateWebsiteReq": {
            "type": "object",
            "required": [
                "description",
                "name",
                "tags",
                "url"
            ],
            "properties": {
                "avatar": {
                    "description": "网站头像",
                    "type": "string",
                    "example": "https://github.githubassets.com/favicons/favicon.svg"
                },
                "description": {
                    "description": "网站描述",
                    "type": "string",
                    "maxLength": 200,
                    "example": "全球最大的代码托管平台"
                },
                "name": {
                    "description": "网站名称",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "GitHub"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "description": "状态 1启用 0禁用",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "tags": {
                    "description": "标签数组",
                    "type": "string",
                    "example": "开发,代码,开源"
                },
                "url": {
                    "description": "网站地址",
                    "type": "string",
                    "example": "https://github.com"
                }
            }
        },
        "dto.GetFaviconReq": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "description": "网站地址",
                    "type": "string",
                    "example": "https://github.com"
                }
            }
        },
...
        "dto.UpdateWebsiteReq": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "网站头像",
                    "type": "string",
                    "example": "https://github.githubassets.com/favicons/favicon.svg"
                },
                "description": {
                    "description": "网站描述",
                    "type": "string",
                    "maxLength": 200,
                    "example": "全球最大的代码托管平台"
                },
                "name": {
                    "description": "网站名称",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "GitHub"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "description": "状态 1启用 0禁用",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "tags": {
                    "description": "标签数组",
                    "type": "string",
                    "example": "开发,代码,开源"
                },
                "url": {
                    "description": "网站地址",
                    "type": "string",
                    "example": "https://github.com"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "vo.GetFaviconVO": {
            "type": "object",
            "properties": {
                "favicon": {
                    "description": "网站图标URL",
                    "type": "string",
                    "example": "https://github.githubassets.com/favicons/favicon.svg"
                }
            }
        },
...
        "vo.WebsiteVO": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "网站头像",
                    "type": "string",
                    "example": "https://github.githubassets.com/favicons/favicon.svg"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "网站描述",
                    "type": "string",
                    "example": "全球最大的代码托管平台"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "name": {
                    "description": "网站名称",
                    "type": "string",
                    "example": "GitHub"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "description": "状态 1启用 0禁用",
                    "type": "integer",
                    "example": 1
                },
                "tags": {
                    "description": "标签数组",
                    "type": "string",
                    "example": "开发,代码,开源"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "url": {
                    "description": "网站地址",
                    "type": "string",
                    "example": "https://github.com"
                }
            }
        }
    }
}